add: session storage

+++ frontend/src/Components/Pages/LoginPage.js
@@ -3,7 +3,6 @@ import Navbar from '../Navbar/Navbar';
 import Navigate from '../Router/Navigate';
 import { clearPage } from '../../utils/render';
 
-
 function renderLoginForm() {
   const main = document.querySelector('main');
   main.innerHTML = `
@@ -68,36 +67,43 @@ function renderLoginForm() {
                 </div>
             </div>
         </div>
-    </div>`
+    </div>`;
 
   const btnRegister = document.getElementById('register');
-  btnRegister.addEventListener('click',handleRegisterClick);
+  btnRegister.addEventListener('click', handleRegisterClick);
 
   const btnLogin = document.getElementById('login');
-  btnLogin.addEventListener('click',handleLoginClick);
+  btnLogin.addEventListener('click', handleLoginClick);
 
+  const souvenir = document.getElementById('rememberMe');
+  souvenir.addEventListener('change', remember);
 
   const passswordInput = document.querySelector('#password');
   const passwordBtn = document.querySelector('#hidePassword');
-  
-  passwordBtn.addEventListener('click',(e)=> {
+
+  passwordBtn.addEventListener('click', (e) => {
     e.preventDefault();
     const passwordVisibility = passswordInput.getAttribute('type');
-    if(passwordVisibility === 'password'){
+    if (passwordVisibility === 'password') {
       passswordInput.setAttribute('type', 'text');
-      passwordBtn.innerHTML = '<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-eye" viewBox="0 0 16 16"><path d="M16 8s-3-5.5-8-5.5S0 8 0 8s3 5.5 8 5.5S16 8 16 8M1.173 8a13.133 13.133 0 0 1 1.66-2.043C4.12 4.668 5.88 3.5 8 3.5c2.12 0 3.879 1.168 5.168 2.457A13.133 13.133 0 0 1 14.828 8c-.058.087-.122.183-.195.288-.335.48-.83 1.12-1.465 1.755C11.879 11.332 10.119 12.5 8 12.5c-2.12 0-3.879-1.168-5.168-2.457A13.134 13.134 0 0 1 1.172 8z"/><path d="M8 5.5a2.5 2.5 0 1 0 0 5 2.5 2.5 0 0 0 0-5M4.5 8a3.5 3.5 0 1 1 7 0 3.5 3.5 0 0 1-7 0"/></svg>'
-  
-    }else{
+      passwordBtn.innerHTML =
+        '<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-eye" viewBox="0 0 16 16"><path d="M16 8s-3-5.5-8-5.5S0 8 0 8s3 5.5 8 5.5S16 8 16 8M1.173 8a13.133 13.133 0 0 1 1.66-2.043C4.12 4.668 5.88 3.5 8 3.5c2.12 0 3.879 1.168 5.168 2.457A13.133 13.133 0 0 1 14.828 8c-.058.087-.122.183-.195.288-.335.48-.83 1.12-1.465 1.755C11.879 11.332 10.119 12.5 8 12.5c-2.12 0-3.879-1.168-5.168-2.457A13.134 13.134 0 0 1 1.172 8z"/><path d="M8 5.5a2.5 2.5 0 1 0 0 5 2.5 2.5 0 0 0 0-5M4.5 8a3.5 3.5 0 1 1 7 0 3.5 3.5 0 0 1-7 0"/></svg>';
+    } else {
       passswordInput.setAttribute('type', 'password');
-      passwordBtn.innerHTML = '<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-eye-slash" viewBox="0 0 16 16"><path d="M13.359 11.238C15.06 9.72 16 8 16 8s-3-5.5-8-5.5a7.028 7.028 0 0 0-2.79.588l.77.771A5.944 5.944 0 0 1 8 3.5c2.12 0 3.879 1.168 5.168 2.457A13.134 13.134 0 0 1 14.828 8c-.058.087-.122.183-.195.288-.335.48-.83 1.12-1.465 1.755-.165.165-.337.328-.517.486z"/><path d="M11.297 9.176a3.5 3.5 0 0 0-4.474-4.474l.823.823a2.5 2.5 0 0 1 2.829 2.829zm-2.943 1.299.822.822a3.5 3.5 0 0 1-4.474-4.474l.823.823a2.5 2.5 0 0 0 2.829 2.829"/><path d="M3.35 5.47c-.18.16-.353.322-.518.487A13.134 13.134 0 0 0 1.172 8l.195.288c.335.48.83 1.12 1.465 1.755C4.121 11.332 5.881 12.5 8 12.5c.716 0 1.39-.133 2.02-.36l.77.772A7.029 7.029 0 0 1 8 13.5C3 13.5 0 8 0 8s.939-1.721 2.641-3.238l.708.709zm10.296 8.884-12-12 .708-.708 12 12-.708.708"/>g</svg>'
+      passwordBtn.innerHTML =
+        '<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-eye-slash" viewBox="0 0 16 16"><path d="M13.359 11.238C15.06 9.72 16 8 16 8s-3-5.5-8-5.5a7.028 7.028 0 0 0-2.79.588l.77.771A5.944 5.944 0 0 1 8 3.5c2.12 0 3.879 1.168 5.168 2.457A13.134 13.134 0 0 1 14.828 8c-.058.087-.122.183-.195.288-.335.48-.83 1.12-1.465 1.755-.165.165-.337.328-.517.486z"/><path d="M11.297 9.176a3.5 3.5 0 0 0-4.474-4.474l.823.823a2.5 2.5 0 0 1 2.829 2.829zm-2.943 1.299.822.822a3.5 3.5 0 0 1-4.474-4.474l.823.823a2.5 2.5 0 0 0 2.829 2.829"/><path d="M3.35 5.47c-.18.16-.353.322-.518.487A13.134 13.134 0 0 0 1.172 8l.195.288c.335.48.83 1.12 1.465 1.755C4.121 11.332 5.881 12.5 8 12.5c.716 0 1.39-.133 2.02-.36l.77.772A7.029 7.029 0 0 1 8 13.5C3 13.5 0 8 0 8s.939-1.721 2.641-3.238l.708.709zm10.296 8.884-12-12 .708-.708 12 12-.708.708"/>g</svg>';
     }
-  
-  })
-
+  });
 }
 
+let isRememberMeChecked;
 
-function  handleRegisterClick(){
+function remember() {
+  isRememberMeChecked = document.getElementById('rememberMe').checked;
+  console.log('se souvenir de moi : ', isRememberMeChecked);
+}
+
+function handleRegisterClick() {
   Navigate('/register');
 }
 
@@ -118,29 +124,34 @@ async function handleLoginClick(e) {
   };
 
   const response = await fetch('http://localhost:3000/users/login', options);
-  
+
   if (!response.ok) {
     Swal.fire({
-      title: "Le pseudo ou le mot de passe est incorrect",
-      icon: "error",
-      showConfirmButton: true
+      title: 'Le pseudo ou le mot de passe est incorrect',
+      icon: 'error',
+      showConfirmButton: true,
     });
-  }else{
-    const responseData = await response.json(); 
-    localStorage.setItem("token", responseData.token);
+  } else {
+    if (isRememberMeChecked) {
+      const responseData = await response.json();
+      localStorage.setItem('token', responseData.token);
+ 
+    } else {
+      const responseData = await response.json();
+      sessionStorage.setItem('token', responseData.token);
+      
+    }
 
     Swal.fire({
-      title: "Connexion réussie!",
-      icon: "success",
+      title: 'Connexion réussie!',
+      icon: 'success',
       timer: 1000,
-      showConfirmButton: false
+      showConfirmButton: false,
     });
-    
+
     Navbar();
     Navigate('/categories');
   }
-
-  
 }
 
 const LoginPage = () => {
